Partial, incomplete fix of game crash during screen lock
Doesn’t crash anymore on lock.
The problem before was that when the screen is locked, orientation
changes, which recreates the game activity, which deleted the game loop
and game thread, which led to null pointer errors.
Problem was partially fixed by storing important parts of the state of
the game (positions, speeds etc.) during the
‘onRetainNonConfigurationInstance()’ function call in game activity,
and restoring them in onCreate().
At first, tried to use xml change to lock the screen orientation, but
that led to weird glitches in canvas.
Game loop continues while screen is locked, but this shouldn’t be too
hard to fix.
Also, forgot to implement restoration of player’s angle, which is why
the player is flat and motionless when the game starts again.