Game pause and resume during screen lock/unlock sort of works now
When the screen locks, the game is successfully paused, but when it
unlocks it resumes, and during the unlock process the game progresses a
bit. It’s not that big of a deal, but I think it could be changed by
moving lots of code around to different functions, e.g. only restarting
game loop during the SurfaceCreated() function, but that might lead to
further problems because of how the order of the lifecycle is.

It might be better to have the game be in a paused state when the
screen unlocks, which I’m looking into now - it looks like I have to
use some new code involving ‘broadcast receivers’ to detect screen
lock/unlock.